Split game-config-slice into helpers, state, and selectors modules
Extract pure helpers (computeIsValid, defaultRoleSlotCount,
roleCountsFromBuckets, roleBucketsFromCounts, recomputeIsValid) into
game-config-helpers.ts, the state interface and initial state into
game-config-state.ts, and the selectors into game-config-selectors.ts.
The createSlice call and default reducer export stay in the slice file,
which re-exports GameConfigState and the selectors for consumers.

Pure refactor with no behavior change; brings the slice from 461 to 323
lines, under the 400-line source-file CI cap.
