+++ src/store/game-config-slice.ts
@@ -1,18 +1,11 @@
 import type { PayloadAction } from "@reduxjs/toolkit";
-import { createSelector, createSlice } from "@reduxjs/toolkit";
-import { sum } from "lodash";
+import { createSlice } from "@reduxjs/toolkit";
 
-import {
-  DEFAULT_GAME_MODE,
-  GAME_MODES,
-  getAdvancedBucketMaxCapacity,
-} from "@/lib/game/modes";
+import { GAME_MODES } from "@/lib/game/modes";
 import type {
   AdvancedRoleBucket,
   ModeConfig,
   ModeConfigField,
-  RoleBucket,
-  SimpleRoleBucket,
   TimerConfig,
 } from "@/lib/types";
 import {
@@ -23,138 +16,13 @@ import {
 } from "@/lib/types";
 import type { GameConfig } from "@/server/types";
 
-function computeIsValid(
-  gameMode: GameMode,
-  playerCount: number,
-  roleConfigMode: RoleConfigMode,
-  roleCounts: Record<string, number>,
-  roleBuckets: RoleBucket[],
-  modeConfig: ModeConfig,
-): boolean {
-  const modeDefinition = GAME_MODES[gameMode];
-  if (roleConfigMode === RoleConfigMode.Default) {
-    return playerCount >= modeDefinition.minPlayers;
-  }
-  if (roleConfigMode === RoleConfigMode.Advanced) {
-    if (roleBuckets.length === 0) return false;
-    const totalPlayerCount = sum(roleBuckets.map((b) => b.playerCount));
-    const required =
-      modeDefinition.resolveRoleSlotsRequired?.(playerCount, modeConfig) ??
-      modeDefinition.roleSlotsRequired?.(playerCount) ??
-      playerCount;
-    if (totalPlayerCount !== required) return false;
-    // Each advanced bucket must have enough draw capacity to fill its playerCount.
-    for (const bucket of roleBuckets) {
-      if (isSimpleRoleBucket(bucket)) continue;
-      if (getAdvancedBucketMaxCapacity(bucket) < bucket.playerCount)
-        return false;
-    }
-    // Aggregate simple bucket counts for mode-specific validation.
-    const simpleCounts: Record<string, number> = {};
-    for (const bucket of roleBuckets) {
-      if (!isSimpleRoleBucket(bucket)) continue;
-      simpleCounts[bucket.roleId] =
-        (simpleCounts[bucket.roleId] ?? 0) + bucket.playerCount;
-    }
-    if (modeDefinition.validateRoleConfig?.(simpleCounts)) return false;
-    return true;
-  }
-  // Custom mode
-  if (modeDefinition.validateRoleConfig?.(roleCounts)) return false;
-  if (modeDefinition.resolveIsValidRoleCount) {
-    return modeDefinition.resolveIsValidRoleCount(
-      playerCount,
-      roleCounts,
-      modeConfig,
-    );
-  }
-  if (modeDefinition.isValidRoleCount) {
-    return modeDefinition.isValidRoleCount(playerCount, roleCounts);
-  }
-  return sum(Object.values(roleCounts)) === playerCount;
-}
-
-/**
- * Returns the number of role slots to pass to `defaultRoleCount` for a given
- * player count and mode config. Accounts for modes that reserve players as
- * non-role owners (e.g. Werewolf Narrator) and for hidden unassigned roles.
- */
-function defaultRoleSlotCount(
-  gameMode: GameMode,
-  playerCount: number,
-  modeConfig: ModeConfig,
-): number {
-  const modeDefinition = GAME_MODES[gameMode];
-  return (
-    modeDefinition.resolveRoleSlotsRequired?.(playerCount, modeConfig) ??
-    modeDefinition.roleSlotsRequired?.(playerCount) ??
-    playerCount
-  );
-}
-
-function roleCountsFromBuckets(buckets: RoleBucket[]): Record<string, number> {
-  const counts: Record<string, number> = {};
-  for (const bucket of buckets) {
-    if (isSimpleRoleBucket(bucket)) {
-      counts[bucket.roleId] = (counts[bucket.roleId] ?? 0) + bucket.playerCount;
-    }
-    // Advanced multi-role buckets don't map to a simple per-role count
-  }
-  return counts;
-}
-
-function roleBucketsFromCounts(
-  roleCounts: Record<string, number>,
-): SimpleRoleBucket[] {
-  return Object.entries(roleCounts)
-    .filter(([, count]) => count > 0)
-    .map(([roleId, count]) => ({ playerCount: count, roleId }));
-}
-
-export interface GameConfigState {
-  gameMode: GameMode;
-  playerCount: number;
-  roleConfigMode: RoleConfigMode;
-  /** Used for Custom mode (exact counts, min === max). */
-  roleCounts: Record<string, number>;
-  /** Used for Default/Custom/Advanced modes. */
-  roleBuckets: RoleBucket[];
-  showConfigToPlayers: boolean;
-  showRolesInPlay: ShowRolesInPlay;
-  timerConfig: TimerConfig;
-  /** Game-mode-specific configuration. Discriminated by `modeConfig.gameMode`. */
-  modeConfig: ModeConfig;
-  isValid: boolean;
-  /** Increments on every user-initiated action. Used to detect when a sync is needed. */
-  syncVersion: number;
-}
-
-const defaultMode = DEFAULT_GAME_MODE;
-
-const initialState: GameConfigState = {
-  gameMode: defaultMode,
-  playerCount: 5,
-  roleConfigMode: RoleConfigMode.Default,
-  roleCounts: {},
-  roleBuckets: [],
-  showConfigToPlayers: false,
-  showRolesInPlay: ShowRolesInPlay.None,
-  timerConfig: GAME_MODES[defaultMode].defaultTimerConfig,
-  modeConfig: GAME_MODES[defaultMode].defaultModeConfig,
-  isValid: false,
-  syncVersion: 0,
-};
-
-function recomputeIsValid(state: GameConfigState): boolean {
-  return computeIsValid(
-    state.gameMode,
-    state.playerCount,
-    state.roleConfigMode,
-    state.roleCounts,
-    state.roleBuckets,
-    state.modeConfig,
-  );
-}
+import {
+  defaultRoleSlotCount,
+  recomputeIsValid,
+  roleBucketsFromCounts,
+  roleCountsFromBuckets,
+} from "./game-config-helpers";
+import { initialState } from "./game-config-state";
 
 const gameConfigSlice = createSlice({
   name: "gameConfig",
@@ -427,14 +295,8 @@ const gameConfigSlice = createSlice({
   },
 });
 
-export const selectRoleBuckets = (state: GameConfigState): RoleBucket[] =>
-  state.roleBuckets;
-
-// Kept for backward compat with any callers that still reference it — returns empty array now.
-export const selectRoleSlots = createSelector(
-  (state: GameConfigState) => state.roleBuckets,
-  (): never[] => [],
-);
+export { selectRoleBuckets, selectRoleSlots } from "./game-config-selectors";
+export type { GameConfigState } from "./game-config-state";
 
 export const {
   loadConfig,
